tests/wpt/mozilla/tests/shadow-dom/getrootnode-in-deeply-nested-shadow.html

<!--https://github.com/servo/servo/issues/35724 -->
<html>
    <head>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
<body>
    <script>
        'use strict';
        test(() => {
            let original_host = document.createElement("div");
            original_host.id = "original-host";
            let current_host = original_host;
            for (let i = 0; i < 10000; i++) {
                let root = current_host.attachShadow({ mode: "open" });
                let new_host = document.createElement("div");
                root.appendChild(new_host);
                current_host = new_host;
            }

            let root_of_tree = current_host.getRootNode({ composed: true });
            assert_equals(root_of_tree.id, "original-host");
        }, "Calling Node.getRootNode in a deeply nested shadow tree must return the correct result without crashing");
    </script>
</body>
</html>
